pyproject.toml
[project]
name = "uvscem"
dynamic = ["version"]
description = ""
authors = [{ name = "Arne Fahrenwalde", email = "arne@fahrenwal.de" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
	"json5>=0.9.17,<1.0.0",
	"requests>=2.31.0,<3.0.0",
	"dependency-algorithm>=0.1,<0.2",
]

[project.urls]
Repository = "https://github.com/macgeneral/uVSCEM/"

[project.scripts]
uvscem = "uvscem.extension_manager:main"

[dependency-groups]
dev = [
	"ruff>=0.2.2",
	"ty>=0.0.18",
	"pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build.targets.wheel]
packages = ["src/uvscem"]

[tool.uv-dynamic-versioning]
vcs = "git"
style = "semver"
fallback-version = "0.1.3"

[tool.ruff]
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["F", "I"]

[tool.ty.src]
include = ["src"]

[tool.ty.environment]
root = ["./src"]
python-version = "3.10"

[tool.ty.rules]
all = "error"

[tool.pytest.ini_options]
addopts = "-x --cov=src/uvscem --cov-branch --cov-fail-under=100 --cov-report=term-missing:skip-covered"
